:art: Split constants to a seperate file
