+++ lib/constants.js
@@ -1,0 +1,6 @@
+/* @flow */
+
+export const $version = '__$sb_linter_version'
+export const $activated = '__$sb_linter_activated'
+export const $requestLatest = '__$sb_linter_request_latest'
+export const $requestLastReceived = '__$sb_linter_request_last_received'
