+++ lib/helpers.js
@@ -3,12 +3,9 @@
 import arrayUnique from 'lodash.uniq'
 import { Disposable, Range, Point } from 'atom'
 import type { TextEditor } from 'atom'
-import type { Linter, Message, MessageLegacy } from './types'
 
-export const $version = '__$sb_linter_version'
-export const $activated = '__$sb_linter_activated'
-export const $requestLatest = '__$sb_linter_request_latest'
-export const $requestLastReceived = '__$sb_linter_request_last_received'
+import { $version } from './constants'
+import type { Linter, Message, MessageLegacy } from './types'
 
 export function shouldTriggerLinter(
   linter: Linter,
