+++ lib/index.js
@@ -3,6 +3,7 @@
 import { CompositeDisposable, Disposable } from 'atom'
 
 import Linter from './main'
+import { $version } from './constants'
 import type { UI, Linter as LinterProvider, State } from './types'
 
 // Internal variables
@@ -75,6 +76,7 @@ export default {
   serialize(): Object {
     instance.registryMessagesInit()
     return {
+      linters: instance.registryLinters.getProviders().filter(p => p[$version] === 2).map(p => p.name),
       messages: instance.registryMessages.serialize(),
     }
   },
