+++ lib/linter-registry.js
@@ -6,7 +6,7 @@ import type { TextEditor, Disposable } from 'atom'
 
 import * as Helpers from './helpers'
 import * as Validate from './validate'
-import { $version, $activated, $requestLatest, $requestLastReceived } from './helpers'
+import { $version, $activated, $requestLatest, $requestLastReceived } from './constants'
 import type { Linter } from './types'
 
 class LinterRegistry {
