+++ lib/message-registry.js
@@ -3,7 +3,8 @@
 import { CompositeDisposable, Emitter } from 'atom'
 import debounce from 'sb-debounce'
 import type { Disposable, TextBuffer } from 'atom'
-import { $version, trimMessages, messageKey, messageKeyLegacy } from './helpers'
+import { $version } from './constants'
+import { trimMessages, messageKey, messageKeyLegacy } from './helpers'
 import type { MessagesPatch, Message, MessageLegacy, Linter, Linter$Message$Map, State } from './types'
 
 class MessageRegistry {
