+++ lib/types.js
@@ -106,5 +106,6 @@ export type Linter$Message$Map = {
 }
 
 export type State = {
+  linters?: Array<string>,
   messages?: Linter$Message$Map,
 }
